refactor: (#1) 버튼 태그의 기본 속성 활용하도록 변경
